Register poupData as global state
